backend/requirements.txt
anthropic
pandas
requests
flask
flask-cors
python-dotenv
